backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic
pymupdf4llm
langchain_huggingface
langchain_core
bcrypt
requests
